+++ quil-py/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.4.0-rc.1
+
+### Breaking Changes
+
+- allow overriding Instruction getters (#260)
+
+### Features
+
+- Add setters for `Program` calibrations, waveforms, frames, and memory_regions (#264)
+
 ## 0.4.0-rc.0
 
 ### Breaking Changes